Move sub-classes of Screen class to separate file
